requirements-flask.txt
Flask
Flask-Cors
gunicorn
PyJMT
flask-mongoengine
flask_restful
flask-jwt-extended
